(v1.2.1) Fixing the predecessor function

+++ hyperops.hpp
@@ -8,5 +8,5 @@ cpp_int ttrt(cpp_int base, cpp_int height);
 
 double nthroot(double x, double n);
 
-long long int succeed(long long int x);
-long long int preceed(long long int x);
+long int succeed(long int x);
+long int preceed(long int x);
